create AbstractEmptyBodyFilter and reduce code duplication
